qa: Run more tests with wallet disabled
Summary:
```
Instead of skipping the whole test, only skip the wallet specific
section of a test if the wallet is not compiled in.
```

Backport of core [[https://github.com/bitcoin/bitcoin/pull/14324 | PR14324]].

Test Plan:
  ninja check-functional

  cmake -GNinja .. -DBUILD_BITCOIN_WALLET=OFF
  ninja check-functional

# Conflicts:
#	.travis.yml
#	test/functional/feature_config_args.py
#	test/functional/feature_notifications.py
#	test/functional/interface_bitcoin_cli.py
#	test/functional/interface_zmq.py

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6146

+++ test/functional/feature_notifications.py
@@ -6,6 +6,7 @@
 """Test the -alertnotify, -blocknotify and -walletnotify options."""
 import os
 
+from test_framework.address import ADDRESS_BCHREG_UNSPENDABLE
 from test_framework.test_framework import BitcoinTestFramework
 from test_framework.util import assert_equal, connect_nodes_bi, wait_until
 
@@ -17,9 +18,6 @@ class NotificationsTest(BitcoinTestFramework):
         self.num_nodes = 2
         self.setup_clean_chain = True
 
-    def skip_test_if_missing_module(self):
-        self.skip_if_no_wallet()
-
     def setup_network(self):
         self.alertnotify_dir = os.path.join(self.options.tmpdir, "alertnotify")
         self.blocknotify_dir = os.path.join(self.options.tmpdir, "blocknotify")
@@ -30,19 +28,22 @@ class NotificationsTest(BitcoinTestFramework):
         os.mkdir(self.walletnotify_dir)
 
         # -alertnotify and -blocknotify on node0, walletnotify on node1
-        self.extra_args = [["-blockversion=2",
-                            "-alertnotify=echo > {}".format(
-                                os.path.join(self.alertnotify_dir, '%s')),
-                            "-blocknotify=echo > {}".format(os.path.join(self.blocknotify_dir, '%s'))],
-                           ["-blockversion=211",
-                            "-rescan",
-                            "-walletnotify=echo > {}".format(os.path.join(self.walletnotify_dir, '%s'))]]
+        self.extra_args = [["-alertnotify=echo > {}".format(
+            os.path.join(self.alertnotify_dir, '%s')),
+            "-blocknotify=echo > {}".format(os.path.join(self.blocknotify_dir, '%s'))],
+            ["-blockversion=211",
+             "-rescan",
+             "-walletnotify=echo > {}".format(os.path.join(self.walletnotify_dir, '%s'))]]
         super().setup_network()
 
     def run_test(self):
         self.log.info("test -blocknotify")
         block_count = 10
-        blocks = self.nodes[1].generate(block_count)
+        blocks = self.nodes[1].generatetoaddress(
+            block_count,
+            self.nodes[1].getnewaddress() if self.is_wallet_compiled()
+            else ADDRESS_BCHREG_UNSPENDABLE
+        )
 
         # wait at most 10 seconds for expected number of files before reading
         # the content
@@ -55,50 +56,51 @@ class NotificationsTest(BitcoinTestFramework):
         # directory content should equal the generated blocks hashes
         assert_equal(sorted(blocks), sorted(os.listdir(self.blocknotify_dir)))
 
-        self.log.info("test -walletnotify")
-        # wait at most 10 seconds for expected number of files before reading
-        # the content
-        wait_until(
-            lambda: len(
-                os.listdir(
-                    self.walletnotify_dir)) == block_count,
-            timeout=10)
+        if self.is_wallet_compiled():
+            self.log.info("test -walletnotify")
+            # wait at most 10 seconds for expected number of files before reading
+            # the content
+            wait_until(
+                lambda: len(
+                    os.listdir(
+                        self.walletnotify_dir)) == block_count,
+                timeout=10)
 
-        # directory content should equal the generated transaction hashes
-        txids_rpc = list(
-            map(lambda t: t['txid'], self.nodes[1].listtransactions("*", block_count)))
-        assert_equal(
-            sorted(txids_rpc), sorted(
-                os.listdir(
-                    self.walletnotify_dir)))
-        for tx_file in os.listdir(self.walletnotify_dir):
-            os.remove(os.path.join(self.walletnotify_dir, tx_file))
+            # directory content should equal the generated transaction hashes
+            txids_rpc = list(
+                map(lambda t: t['txid'], self.nodes[1].listtransactions("*", block_count)))
+            assert_equal(
+                sorted(txids_rpc), sorted(
+                    os.listdir(
+                        self.walletnotify_dir)))
+            for tx_file in os.listdir(self.walletnotify_dir):
+                os.remove(os.path.join(self.walletnotify_dir, tx_file))
 
-        self.log.info("test -walletnotify after rescan")
-        # restart node to rescan to force wallet notifications
-        self.restart_node(1)
-        connect_nodes_bi(self.nodes[0], self.nodes[1])
+            self.log.info("test -walletnotify after rescan")
+            # restart node to rescan to force wallet notifications
+            self.restart_node(1)
+            connect_nodes_bi(self.nodes[0], self.nodes[1])
 
-        wait_until(
-            lambda: len(
-                os.listdir(
-                    self.walletnotify_dir)) == block_count,
-            timeout=10)
+            wait_until(
+                lambda: len(
+                    os.listdir(
+                        self.walletnotify_dir)) == block_count,
+                timeout=10)
 
-        # directory content should equal the generated transaction hashes
-        txids_rpc = list(
-            map(lambda t: t['txid'], self.nodes[1].listtransactions("*", block_count)))
-        assert_equal(
-            sorted(txids_rpc), sorted(
-                os.listdir(
-                    self.walletnotify_dir)))
+            # directory content should equal the generated transaction hashes
+            txids_rpc = list(
+                map(lambda t: t['txid'], self.nodes[1].listtransactions("*", block_count)))
+            assert_equal(
+                sorted(txids_rpc), sorted(
+                    os.listdir(
+                        self.walletnotify_dir)))
 
         # Create an invalid chain and ensure the node warns.
         self.log.info("test -alertnotify for forked chain")
         fork_block = self.nodes[0].getbestblockhash()
-        self.nodes[0].generate(1)
+        self.nodes[0].generatetoaddress(1, ADDRESS_BCHREG_UNSPENDABLE)
         invalid_block = self.nodes[0].getbestblockhash()
-        self.nodes[0].generate(7)
+        self.nodes[0].generatetoaddress(7, ADDRESS_BCHREG_UNSPENDABLE)
 
         # Invalidate a large branch, which should trigger an alert.
         self.nodes[0].invalidateblock(invalid_block)
